Use `visibility: hidden` instead of `display: none` to hide settings icon.
Messenger relies on icon’s `offsetParent` to determine whether to show the popover menu. Or, at least, add or remove the `hidden_elem` CSS class to its element.

When element has `display: none` its `offsetParent` is null. That means the we couldn’t rely on `hidden_elem` class to check which popover is “active”.